Add ContactLabel model & attribs to Contact model
ContactLabel is available to be parsed in responses for personal,
corporate, and alliance contacts. A model object is added to capture
this data and exposed through the ContactListParser.

Simiarly two attributes which were available in all contacts have
been added to the model and the values parsed from the APIresponse.

Fixes: #52
Fixes: #53